refactor: split gui and recorder
