+++ gui/__init__.py
@@ -1,0 +1,2 @@
+from .main_window import run_gui, ScribeApp
+from .dialogs import SettingsDialog
